+++ build.gradle
@@ -15,8 +15,8 @@ buildscript {
     }
 }
 
-group = 'RS2-Cache-Library'
-version = '1.0-SNAPSHOT'
+group = 'RS-Cache-Library'
+version = '6.0'
 
 description = ""
 
@@ -42,5 +42,7 @@ compileTestKotlin {
 dependencies {
     implementation group: 'com.github.jponge', name: 'lzma-java', version:'1.3'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+    implementation 'org.apache.ant:ant:1.10.7'
+    implementation 'com.displee:disio:1.3'
     testCompile group: 'junit', name: 'junit', version:'4.12'
 }
