--- src/main/java/org/displee/progress/ProgressListener.java
@@ -1,9 +1,0 @@
-package org.displee.progress;
-
-public interface ProgressListener {
-
-    void notify(double progress, String message);
-
-    void finish(String title, String message);
-
-}
