+++ src/main/kotlin/com/displee/cache/ProgressListener.kt
@@ -1,0 +1,6 @@
+package com.displee.cache
+
+interface ProgressListener {
+    fun notify(progress: Double, message: String?)
+    fun finish(title: String?, message: String?)
+}
