+++ src/main/kotlin/com/displee/compress/CompressionType.kt
@@ -1,0 +1,12 @@
+package com.displee.compress
+
+enum class CompressionType {
+    NONE,
+    BZIP2,
+    GZIP,
+    LZMA;
+
+    companion object {
+        val compressionTypes = values()
+    }
+}
